fix: rename types
