+++ core/transaction-decorator/index.ts
@@ -1,4 +1,4 @@
-export * from '@eggjs/tegg-types/transaction-decorator';
+export * from '@eggjs/tegg-types/transaction';
 export * from './src/decorator/Transactional';
 export * from './src/builder/TransactionMetaBuilder';
 export * from './src/util/TransactionMetadataUtil';
