+++ core/types/dal/enum/ColumnFormat.ts
