+++ core/types/dal/enum/ColumnType.ts
