+++ core/types/dal/enum/CompressionType.ts
