+++ core/types/dal/enum/IndexStoreType.ts
