+++ core/types/dal/enum/IndexType.ts
