+++ core/types/dal/enum/InsertMethod.ts
