+++ core/types/dal/enum/RowFormat.ts
