+++ core/types/dal/enum/SqlType.ts
@@ -1,0 +1,7 @@
+export enum SqlType {
+  BLOCK = 'BLOCK',
+  INSERT = 'INSERT',
+  SELECT = 'SELECT',
+  UPDATE = 'UPDATE',
+  DELETE = 'DELETE',
+}
