+++ core/types/dal/enum/Templates.ts
@@ -1,0 +1,5 @@
+export enum Templates {
+  BASE_DAO = 'base_dao',
+  DAO = 'dao',
+  EXTENSION = 'extension',
+}
