+++ core/types/dal/type/CodeGenerator.ts
@@ -1,0 +1,6 @@
+export interface CodeGeneratorOptions {
+  moduleDir: string;
+  moduleName: string;
+  teggPkg?: string;
+  dalPkg?: string;
+}
