+++ core/types/dal/type/Spatial.ts
