+++ core/types/index.ts
@@ -1,13 +1,12 @@
-export * from './aop-decorator';
+export * from './aop';
 export * from './common';
 export * from './controller-decorator';
 export * from './core-decorator';
-export * from './dal-decorator';
-export * from './dal-runtime';
+export * from './dal';
 export * from './dynamic-inject';
 export * from './lifecycle';
 export * from './metadata';
-export * from './orm-decorator';
+export * from './orm';
 export * from './runtime';
-export * from './schedule-decorator';
-export * from './transaction-decorator';
+export * from './schedule';
+export * from './transaction';
