+++ core/types/schedule.ts
